refactor: move UpdateUserPreference to user-preference folder
