+++ src/domain/contracts/user-preference/index.ts
@@ -1,1 +1,2 @@
 export * from './add-user-preference'
+export * from './update-user-preference'
