+++ src/domain/contracts/user-preference/update-user-preference.ts
@@ -1,9 +1,9 @@
-import { type NonExistentUserPreferenceError } from '@/domain/errors'
-import { type ActiveType, type Frequency } from '@/domain/models'
-import { type Either } from '@/shared/either'
+import type { NonExistentUserPreferenceError } from '@/domain/errors'
+import type { ActiveType, Frequency } from '@/domain/models'
+import type { Either } from '@/shared/either'
 
 export interface UpdateUserPreferenceData {
-  id: string
+  userId: string
   frequency?: Frequency
   activeType?: ActiveType
 }
