+++ src/domain/contracts/user/index.ts
@@ -1,4 +1,3 @@
 export * from './add-user'
 export * from './auth'
-export * from './update-user-preferences'
 export * from './update-user'
