+++ src/infra/db/prisma/data/user-preference/update-user-preference/update-user-preference-prisma-repo.ts
@@ -1,9 +1,8 @@
-import { type UpdateUserPreferenceData } from '@/domain/contracts/user'
-import { type UpdateUserPreferenceRepo } from '@/usecases/contracts/db/user'
+import type { UpdateUserPreferenceRepo, UpdateUserPreferenceRepoData } from '@/usecases/contracts/db/user-preference'
 import { PrismaHelper } from '@/infra/db/prisma/helpers/prisma-helper'
 
 export class UpdateUserPreferencePrismaRepo implements UpdateUserPreferenceRepo {
-  async execute (data: UpdateUserPreferenceData): Promise<void> {
+  async execute (data: UpdateUserPreferenceRepoData): Promise<void> {
     const prisma = await PrismaHelper.getPrisma()
     const { id } = data
     await prisma.userPreference.update({ where: { id }, data })
