+++ src/main/factories/infra/db/prisma/user-preference/update-user-preference-prisma-repo-factory.ts
@@ -1,6 +1,6 @@
 import { UpdateUserPreferencePrismaRepo } from '@/infra/db/prisma/data/user-preference/update-user-preference/update-user-preference-prisma-repo'
-import { type UpdateUserPreferenceRepo } from '@/usecases/contracts/db/user'
+import { type UpdateUserPreferenceRepo } from '@/usecases/contracts/db/user-preference'
 
-export const makeUpdateUserPreferenceRepo = (): UpdateUserPreferenceRepo => {
+export const makeUpdateUserPreferencePrismaRepo = (): UpdateUserPreferenceRepo => {
   return new UpdateUserPreferencePrismaRepo()
 }
