+++ src/main/factories/usecases/user-preference/update-user-preference-usecase-factory.ts
@@ -1,11 +1,11 @@
-import { type UpdateUserPreference } from '@/domain/contracts/user'
+import { type UpdateUserPreference } from '@/domain/contracts/user-preference'
 import { UpdateUserPreferenceUseCase } from '@/usecases/implementations/user-preference/update-user-preference/update-user-preference-usecase'
 import { makeFindUserPreferenceByIdPrismaRepo } from '../../infra/db/prisma/user-preference/find-preference-by-id-prisma-repo-factory'
-import { makeUpdateUserPreferenceRepo } from '../../infra/db/prisma/user/update-user-preference-repo-factory'
+import { makeUpdateUserPreferencePrismaRepo } from '../../infra/db/prisma/user-preference/update-user-preference-prisma-repo-factory'
 
 export const makeUpdateUserPreferenceUsecase = (): UpdateUserPreference => {
   return new UpdateUserPreferenceUseCase(
     makeFindUserPreferenceByIdPrismaRepo(),
-    makeUpdateUserPreferenceRepo()
+    makeUpdateUserPreferencePrismaRepo()
   )
 }
