+++ src/usecases/contracts/db/user-preference/index.ts
@@ -1,2 +1,3 @@
 export * from './find-user-preference-by-id-repo'
 export * from './add-user-preference-repo'
+export * from './update-user-preference-repo'
