+++ src/usecases/contracts/db/user-preference/update-user-preference-repo.ts
@@ -1,0 +1,11 @@
+import type { ActiveType, Frequency } from '@/domain/models'
+
+export type UpdateUserPreferenceRepoData = {
+  id: string
+  frequency?: Frequency
+  activeType?: ActiveType
+}
+
+export interface UpdateUserPreferenceRepo {
+  execute: (data: UpdateUserPreferenceRepoData) => Promise<void>
+}
