+++ src/usecases/contracts/db/user/index.ts
@@ -1,4 +1,3 @@
 export * from './find-user-by-email-repo'
 export * from './add-user-repo'
-export * from './update-user-preference-repo'
 export * from './update-user-repo'
