--- src/usecases/contracts/db/user/update-user-preference-repo.ts
@@ -1,5 +1,0 @@
-import { type UpdateUserPreferenceData } from '@/domain/contracts/user'
-
-export interface UpdateUserPreferenceRepo {
-  execute: (data: UpdateUserPreferenceData) => Promise<void>
-}
