+++ src/usecases/implementations/user-preference/update-user-preference/update-user-preference-usecase.spec.ts
@@ -1,9 +1,8 @@
-import { type UpdateUserPreferenceData } from '@/domain/contracts/user'
 import { NonExistentUserPreferenceError } from '@/domain/errors'
 import { type UserPreferenceModel } from '@/domain/models'
-import { type UpdateUserPreferenceRepo } from '@/usecases/contracts/db/user'
 import { UpdateUserPreferenceUseCase } from './update-user-preference-usecase'
-import { type FindUserPreferenceByIdRepo } from '@/usecases/contracts/db/user-preference'
+import { type UpdateUserPreferenceRepo, type FindUserPreferenceByIdRepo, type UpdateUserPreferenceRepoData } from '@/usecases/contracts/db/user-preference'
+import { type UpdateUserPreferenceData } from '@/domain/contracts/user-preference'
 
 const makeFakeUserPreferenceModel = (): UserPreferenceModel => ({
   id: 'any_user_id',
@@ -12,8 +11,9 @@ const makeFakeUserPreferenceModel = (): UserPreferenceModel => ({
 })
 
 const makeFakeUpdateUserPreferenceData = (): UpdateUserPreferenceData => ({
-  id: 'any_user_id',
-  frequency: 'daily'
+  userId: 'any_user_id',
+  frequency: 'daily',
+  activeType: 'gameMaster'
 })
 
 const makeFindUserPreferenceByIdRepo = (): FindUserPreferenceByIdRepo => {
@@ -27,7 +27,7 @@ const makeFindUserPreferenceByIdRepo = (): FindUserPreferenceByIdRepo => {
 
 const makeUpdateUserPreferenceRepo = (): UpdateUserPreferenceRepo => {
   class UpdateUserPreferenceRepoStub implements UpdateUserPreferenceRepo {
-    async execute (data: UpdateUserPreferenceData): Promise<void> {
+    async execute (data: UpdateUserPreferenceRepoData): Promise<void> {
       await Promise.resolve()
     }
   }
@@ -68,7 +68,7 @@ describe('UpdateUserPreferenceUsecase', () => {
     const { sut, updateUserPreferenceRepoStub } = makeSut()
     const executeSpy = jest.spyOn(updateUserPreferenceRepoStub, 'execute')
     await sut.perform(makeFakeUpdateUserPreferenceData())
-    expect(executeSpy).toHaveBeenCalledWith(makeFakeUpdateUserPreferenceData())
+    expect(executeSpy).toHaveBeenCalledWith(makeFakeUserPreferenceModel())
   })
 
   it('Should throw if UpdateUserPreferenceRepo throws', async () => {
