+++ src/usecases/implementations/user-preference/update-user-preference/update-user-preference-usecase.ts
@@ -1,8 +1,7 @@
-import type { UpdateUserPreferenceResponse, UpdateUserPreference, UpdateUserPreferenceData } from '@/domain/contracts/user'
+import type { UpdateUserPreference, UpdateUserPreferenceData, UpdateUserPreferenceResponse } from '@/domain/contracts/user-preference'
 import { NonExistentUserPreferenceError } from '@/domain/errors'
 import { left, right } from '@/shared/either'
-import { type UpdateUserPreferenceRepo } from '@/usecases/contracts/db/user'
-import { type FindUserPreferenceByIdRepo } from '@/usecases/contracts/db/user-preference'
+import type { FindUserPreferenceByIdRepo, UpdateUserPreferenceRepo } from '@/usecases/contracts/db/user-preference'
 
 export class UpdateUserPreferenceUseCase implements UpdateUserPreference {
   constructor (
@@ -11,12 +10,12 @@ export class UpdateUserPreferenceUseCase implements UpdateUserPreference {
   ) {}
 
   async perform (data: UpdateUserPreferenceData): Promise<UpdateUserPreferenceResponse> {
-    const { id } = data
+    const { userId: id, ...otherData } = data
     const userPreferenceExists = await this.findUserPreferenceByIdRepo.execute(id)
     if (!userPreferenceExists) {
       return left(new NonExistentUserPreferenceError(id))
     }
-    await this.updateUserPreferenceRepo.execute(data)
+    await this.updateUserPreferenceRepo.execute({ id, ...otherData })
     return right(null)
   }
 }
